requirements.txt
fastapi>=0.111.0
uvicorn
psycopg2-binary
pydantic-settings